Files added index_code.rb and check_errors.rb
